requirements.txt

---

# 📦 requirements.txt

```txt
# Core app
streamlit==1.49.1
streamlit-autorefresh==1.0.1
python-dotenv==1.1.1

# LangChain ecosystem
langchain==0.3.27
langgraph==0.6.7
langchain-openai==0.2.14
langchain-anthropic==0.3.19

# Providers
openai==1.99.1
anthropic==0.67.0

# Data & memory
sqlite3-binary==0.0.1  # if SQLite bindings are needed
pandas>=2.2.0
numpy>=1.26.0

# Utilities
blinker>=1.9.0
watchdog>=6.0.0
python-decouple>=3.8
requests>=2.32.4
pillow>=11.3.0
protobuf>=5.29.5
pyarrow>=7.0
pydeck>=0.9.1
tenacity>=9.1.2
toml>=0.10.2
tornado>=6.5.2
altair>=5.5.0
